refactor(activerecord): relocate describeIfMysqlAdapter out of the adapters tree
